style(native): apply clippy::all autofixes (machine-applicable)
cargo clippy --fix on the workspace (pedantic suppressed) — 12 files of
machine-applicable clippy::all suggestions (needless_range_loop, map_or,
explicit-iter, etc.). Verified: cargo check passes, and the touched
crates' own tests pass (incl. ol_fountain property_roundtrip, so PRNG /
fountain determinism is preserved — these are behavior-preserving
transforms). Remaining clippy::all warnings need manual fixes; the
~581 clippy::pedantic warnings (mostly doc-backticks) are opt-in style,
to be addressed via config when the native lint gate lands.

+++ native/ol_fountain/src/rng.rs
@@ -66,7 +66,7 @@ impl SplitMix64 {
     pub fn next_u32_below(&mut self, range: u32) -> u32 {
         debug_assert!(range > 0);
         let r = u64::from(range);
-        let x = (self.next_u64() >> 32) as u64; // u32 reduced
+        let x = (self.next_u64() >> 32); // u32 reduced
         ((x.wrapping_mul(r)) >> 32) as u32
     }
